[#18][FRONT, BACK][도서 목록 조회 및 검색] - 엘라스틱 서치 구현 (100%) (#377)
* feat(엘라스틱): 도서 등록 로직 수정

# Conflicts:
#	backend-server/src/main/resources/application-local.yml

* feat(엘라스틱): 엘라스틱 setting 및 mapping 설정

* feat(엘라스틱): 도서 검색 기능 추가

* feat(엘라스틱): 도서 수정 로직 진행중

* feat(엘라스틱): 도서 수정시 엘라스틱 정보 수정

* feat(엘라스틱): 엘라스틱 서치 백엔드 구현 완료 (95% 리뷰 등록 부분 수정만 남음)

* feat(엘라스틱): FE 연결 완료

* feat(엘라스틱): FE 연결 완료

* feat(엘라스틱): 구현 완료

* fix(test): test 수정 및 프론트 관리자페이지 사이드바 수정

* fix(FE): 불필요한 코드 삭제

+++ backend-server/src/main/java/shop/bluebooktle/backend/book/service/impl/AuthorServiceImpl.java
@@ -1,5 +1,7 @@
 package shop.bluebooktle.backend.book.service.impl;
 
+import java.util.List;
+
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -7,15 +9,18 @@ import org.springframework.transaction.annotation.Transactional;
 
 import lombok.RequiredArgsConstructor;
 import shop.bluebooktle.backend.book.entity.Author;
+import shop.bluebooktle.backend.book.entity.Book;
+import shop.bluebooktle.backend.book.entity.BookAuthor;
 import shop.bluebooktle.backend.book.repository.AuthorRepository;
+import shop.bluebooktle.backend.book.repository.BookAuthorRepository;
 import shop.bluebooktle.backend.book.service.AuthorService;
+import shop.bluebooktle.backend.elasticsearch.service.BookElasticSearchService;
 import shop.bluebooktle.common.dto.book.request.author.AuthorRegisterRequest;
 import shop.bluebooktle.common.dto.book.request.author.AuthorUpdateRequest;
 import shop.bluebooktle.common.dto.book.response.author.AuthorResponse;
 import shop.bluebooktle.common.exception.book.AuthorAlreadyExistsException;
-import shop.bluebooktle.common.exception.book.AuthorFieldNullException;
+import shop.bluebooktle.common.exception.book.AuthorCannotDeleteException;
 import shop.bluebooktle.common.exception.book.AuthorNotFoundException;
-import shop.bluebooktle.common.exception.book.AuthorUpdateFieldMissingException;
 
 @Service
 @Transactional
@@ -23,19 +28,13 @@ import shop.bluebooktle.common.exception.book.AuthorUpdateFieldMissingException;
 public class AuthorServiceImpl implements AuthorService {
 
 	private final AuthorRepository authorRepository;
+	private final BookAuthorRepository bookAuthorRepository;
+	private final BookElasticSearchService bookElasticSearchService;
 
 	@Override
 	public void registerAuthor(AuthorRegisterRequest authorRegisterRequest) {
 
 		String name = authorRegisterRequest.getName();
-
-		if (name == null) {
-			throw new AuthorFieldNullException();
-		}
-
-		if (name.trim().isEmpty()) {
-			throw new AuthorFieldNullException();
-		}
 
 		if (authorRepository.existsByName(name)) {
 			throw new AuthorAlreadyExistsException(name);
@@ -54,7 +53,7 @@ public class AuthorServiceImpl implements AuthorService {
 	public AuthorResponse getAuthor(Long authorId) {
 
 		Author author = authorRepository.findById(authorId)
-			.orElseThrow(() -> new AuthorNotFoundException(authorId)); // #TODO
+			.orElseThrow(() -> new AuthorNotFoundException(authorId));
 
 		return AuthorResponse.builder()
 			.id(author.getId())
@@ -67,15 +66,21 @@ public class AuthorServiceImpl implements AuthorService {
 	public void updateAuthor(Long authorId, AuthorUpdateRequest authorUpdateRequest) {
 
 		String name = authorUpdateRequest.getName();
-
-		if ((name == null || name.trim().isEmpty())) {
-			throw new AuthorUpdateFieldMissingException(); // #TODO
-		}
-
 		Author author = authorRepository.findById(authorId).orElseThrow(() -> new AuthorNotFoundException(authorId));
 
-		author.setName(name);
+		if (authorRepository.existsByName(name)) {
+			throw new AuthorAlreadyExistsException(name);
+		}
 
+		List<Book> bookList = bookAuthorRepository.findByAuthor(author)
+			.stream()
+			.map(BookAuthor::getBook)
+			.toList();
+
+		// 엘라스틱에 등록된 도서 정보 수정
+		bookElasticSearchService.updateAuthorName(bookList, name, author.getName());
+
+		author.setName(name);
 		authorRepository.save(author);
 
 	}
@@ -84,6 +89,11 @@ public class AuthorServiceImpl implements AuthorService {
 	public void deleteAuthor(Long authorId) {
 
 		Author author = authorRepository.findById(authorId).orElseThrow(() -> new AuthorNotFoundException(authorId));
+
+		// 출판사에 등록된 도서가 있을 경우 삭제 불가
+		if (bookAuthorRepository.existsByAuthor(author)) {
+			throw new AuthorCannotDeleteException();
+		}
 		authorRepository.delete(author);
 	}
